peer: instantiate from a parent bus
The bus object contains all that is needed to instantiate the peer
(and soon even more so), so rather than passing it all in to
peer_new(), simply pass in the bus itself. This makes the peer
slightly less generic for the sake of convenience, but even in our
tests it appears that this is exactly what we want.

Signed-off-by: Tom Gundersen <teg@jklm.no>