user: force entry<->registry binding
Right now, the UserEntry object requires the caller to link it, but
unlinks itself. This is, because we want to let the caller manage the
RBTree, but the UserEntry to be ref-counted.

Avoid this asymmetric behavior and at least bind each UserEntry to a
registry right from the beginning. This means, an entry always knows
its registry, but whether or not it is linked is a boolean state
controllable by the caller.

While at it, fix coding style:
 - use calloc()
 - initialize CRBNode entries properly
 - use same order in structure-definition as in initialization

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>