user: make user_usage use same registration as the registry
Similar to the registry, set the owning object in the constructor. This
binds the object to the parent right up-front, without modifying the
field during its lifetime.

The only thing that is modified is the CRBNode, depending on whether it
is linked into the tree or not. But the parent pointer stays accessible
regardless.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>