match: refactor the match object
Distinguish between the RuleKeys and the Filter, one is a description
of a match rule, and the other a description of a message.

Also make the match object refcounted, in case a peer installs the same
match more than once.

Lastly, allow a rule object to be created without specifcying the rules
registry context. This will allow the caller to decide which context to
link depending on the parsed match string.

Signed-off-by: Tom Gundersen <teg@jklm.no>