dispatch: rework API slightly
We now require the events to be polled for and the fd to poll on to
be provided to dispatch_file_init(), and these can no longer be
changed, whithout deintializing the dispatch file.

In order for the callback to be invoked for a given event, the event
must be selected, and it can later be deselected.

Once an event has been handled, it should be clared.

Signed-off-by: Tom Gundersen <teg@jklm.no>