reply: add reply tracking object
When a method call that expects a reply is sent to a peer, a reply
slot is installed in the receiving peer, indexed by the sending peer
id and the serial.

When a reply (or error) is sent, rather than looking up the receiving
peer in the global rbtree, the corresponding reply slot is looked up
in the local reply registry and the peer found from there.

This means that replies and errors can only be addressed at unique
names and can only be sent to peers that expect the given reply,
with the given serial from the given sending peer.

Signed-off-by: Tom Gundersen <teg@jklm.no>