match: remove @sender from filter
The sender field is special, as it is not simply the sender taken
from the message header, but rather it should be instantiated to
each of the names owned by the sending peer.

In a follow-up patch, we will keep match registries per name, so
that a given filter can easily be applied to each name owned by
a peer.

Signed-off-by: Tom Gundersen <teg@jklm.no>