match: keep match rules per-sender
When a broadcast is sent, it should be sent to all subscribers to any
of its names. To make this more straight-forward, store the matches
per-name. This also has the benefitial side-effect of reducing the
number of match rules a given message is compared against (as it will
only be compared against match rules matching its names).

There are four classes of match rules:

 * Wildcard:   these are rules that do not specify a sender at all.
	       all messages should be compared to wildcard rules.

 * Peer:       these rules match against a unique name, which is only
               supported for well-known names currently registered on
               the bus.

 * Driver:     the special well-known name org.freedesktop.DBus does
               not have a name entry, but is otherwise treated like
               any other well-known name.

 * Well-known: all other matches are installed per NameEntry, and
               should be matched by the current primary owner of
               a name. The lifetime of these matches are independent
               of the owner of names, and can even be installed
	       before a name has an owner.

Signed-off-by: Tom Gundersen <teg@jklm.no>