sasl: pass back the reply string, rather than requiring a buffer to be passed in
We used to require the caller to pass in a buffer to write the reply into,
but let's simplify the code by simply passing back the reply strings.

This is not a funcitonal change, as all the strings are static, it just lets
the caller be responsible for copying the strings to the buffer. It also
means that the caller no longer needs to overallocate the buffer space.

Signed-off-by: Tom Gundersen <teg@jklm.no>