sasl: do not hide state changes in 'send' calls
We currently provide sasl send() helpers which output a reply and
perform a state change. This is ambiguous, since the state change is
not necessarily bound to the reply that is sent. Furthermore, we do not
*send* anything, but only set the reply buffer.

Avoid using those send helpers and just inline the calls. This makes
the state transitions obvious and explicit. Furthermore, the reply
handling is rather trivial, anyway.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>