socket: keep position up-to-date for messages as well
We have a position marker `data_pos` to remember how far the line
parser went so far. We never really used it for the message parser,
since it is not necessary. However, in the future, we want to be able
to easily figure out whether there is input data left. Hence, keep the
position up-to-date even for messages.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>