socket: split off socket_read()
Right now, the socket output handling is split into kernel I/O and
local queueing. That is, the actual expensive kernel operations are
split of into a separate step in our main-loop iteration. This nicely
encapsulates preemption and buffer, and moves it out of the hot-path.

Unfortunately, input handling was never split into two paths. This was
mainly because we need line and message buffering separately. However,
if we avoid parsing the incoming data, we can still split the data
retrieval into a separate step.

As a first step, we extract socket_read() from the line retrieval
helpers. Follow-ups will then untangle it and integrate it with the
caller loops.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>