error: add error-folding helper
We treat any unhandled error as fatal error. Add the error_fold() helper
to provide a hook for this operation, so debuggers can trace it.

We might also decide to add a debug printf() there. This is still up
for discussion, and will probably be left to the executable that links
the code.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>