driver: introduce driver_goodbye()
This is the conceptual counter-part to Hello(), though it is not
exposed on the bus. It should be called as the last method on a
peer and it cleans up any names, matches and outsanding replies
on the peer. Finally it unregisteres the peer from the bus.

If @silent is true, it causes no notifications or messages to be
generated, and in this case the method cannot fail. This should
only be used when doing the final cleanup of the bus, in which
case all the peers are being disconnected and nobody would be
receiving any notifications anyway.

Signed-off-by: Tom Gundersen <teg@jklm.no>