error: beef up the error helpers to allow full tracing
The error helpers are meant to wrap the return of any fatal error
codes. At the deepest level in our call-stack, error_origin() is called
to start a trace, and to ensure that the error code is non-positive.
A negative error code returned from our own code should always be
fatal, and any fatal (negative or possitive) error code returned
from our own code should be wrapped in error_fold() to get a backtrace
printed to stderr. When error_fold() is called on a negative error code
it ignores the actual error code and just prints a trace, if it is called
on a positive error code it calls into error_origin() to start a new
trace from that point up.

If error_origin() or error_fold() are called on 0, they are no-ops,
though this might be dropped in the future if it does not turn
out to be useful.

Signed-off-by: Tom Gundersen <teg@jklm.no>