tree-wide: update error handling
Hook up the most basic error handling. Note however, that we still
need to go through and make sure all the errors are handled correctly,
we will now fail hard in many cases we should not (but with a nice
backtrace).

peer.c and bus.c were not updated as it's not really clear what we
want from them.

Signed-off-by: Tom Gundersen <teg@jklm.no>