driver: improve error paths
Return DRIVER_E_UNEXPECTED_* in case we are passed a message with
a path/interface/method/signature that is not expected (in which
case an error should be returned to the caller). And return
DRIVER_E_INVALID_MESSAGE in case the message body is invalid
according to the signature. In which case the caller should be
disconnected from the bus.

Signed-off-by: Tom Gundersen <teg@jklm.no>