broker: block signals only during run()
We only dispatch signals during run(), so don't block signals
afterwards. If the surrounding code wants signals blocked as well, then
they better do it manually. However, we should not rely on the manager
to block them for everyone.

We do correctly reset the mask to its previous state, so we are
directly compatible with whoever called this and manages the set
itself.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>