message: implement generic metadata parser
Add a new function called message_parse_metadata(), which parses the
message metadata and returns it in a convenient structure for direct
access.

This is almost identical to what we did so far in peer.c, but has
slightly stricter checks, which were aligned with what dbus-daemon(1)
does.

Note that all of the non-implemented checks have a XXX marker so far.
So this has all been checked with dbus-daemon(1) sources and should be
identical now.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>