tree-wide: adher to coding-style for all embedded structs
Two functions are always provided foobar_init() and foobar_deinit().
The latter always returns void and never takes any additional arguments.

If foobar_init() returns void, then a macro is provided FOOBAR_INIT which
can be assigned to a foobar struct and have the same effect as calling
foobar_init() on it. If FOOBAR_INIT is equivalent to the empty struct,
it is ommitted and the empty struct is used instead.

If FOOBAR_INIT does not exist, or takes arguments (other than the
struct itself) then FOOBAR_NULL should be provided, which takes no
arguments (other than potentially the struct itself) and gives a
struct that foobar_deinit() can be called on. In case FOOBAR_NULL
is equivalent to the empty struct, it is omitted and the empty
struct used instead.

foobar_deinit() should reinitialize the struct to foobar_init() if it
takes no arguments and returns void. Otherwise, it should reinitialize
the struct to FOOBAR_NULL. It should also assert() in case
reinitializing the struct would leak memory, or in other ways it is
detected that cleanup has not been performed.

Signed-off-by: Tom Gundersen <teg@jklm.no>