unique-name: add unique name helpers
This converts between string and uint64_t representation. We reserve
the ID (uint64_t)-1 as UNIQUE_NAME_ID_INVALID. And consider all ids
larger than or equal to UNIQEU_NAME_ID_INVALID as out of range.

There is nothing in the specification disallowing such ids to be
sent to us, but they can never correspond to a real peer, as the
ids are allocated consequtively.

Signed-off-by: Tom Gundersen <teg@jklm.no>