peer: rework queuing of unicast messages
A unicast message can now be queued either on an activatable name or
on a peer. When it is queued on a peer a reply object may be allocated
(if requested), but names no longer have reply objects.

It makes no sense to allocate reply objects already on the name, as
we only know if the method should be finally queued when we know the
receiver. Policy, quotas and conflicting serial numbers may all cause
the final queuing to fail.

Signed-off-by: Tom Gundersen <teg@jklm.no>