socket: select a reasonable value for SOCKET_LINE_PREALLOC
This should hold the longest possible one-sided SASL exchange that
we want to support pipelining. There is no limit to how long SASL
exchanges may be, as the client is allowed to retry on error, but
that really only makes sense if they first receive the error from
the server, in which case we will not be pipelining the whole
exchange.

It is still technically possible for the client to predict the error
and pipeline the retry, but that really makes no sense and is not
something we want to optimize for.

Signed-off-by: Tom Gundersen <teg@jklm.no>