socket: refactor socket_dispatch_read()
In the common case, the input buffer is simply filled with as much
data as possible from the kernel. However, there are two exceptions:

 * when processing lines, if the input buffer is completely full,
   it is reallocated to fit the longest possible line.
 * when processing messages, if the partially read message will not
   fit into the input buffer, the buffer is skipped and the message
   is read directly into the message object.

Reorganize the code a bit to make it clear that these are mutually
exclusive, and that they really only apply in the line and socket
case, respectively.

Signed-off-by: Tom Gundersen <teg@jklm.no>