socket: allow recvmsg() to be limited
Preemption/ratelimiting is essentially implemented by limiting how
much data can be read on every dispatch. However, it is possible
to trigger the input buffer to be resized, essentially arbitrarily
changing the limit for the duration of the connection. Rework this
so that now the size of the input buffer (which needs to potentially
be large for the line reader), is independent of the size of the
reads.

Signed-off-by: Tom Gundersen <teg@jklm.no>