peer: move the metrics from the dispatcher to the peer
Rather than measure the time it takes to dispatch an event,
measure the time it takes to dispatch a single message. This
means we do not account for all the time spent in the daemon,
but this is more accurate as each event can dispatch an (almost)
arbitrary number of messages.

Signed-off-by: Tom Gundersen <teg@jklm.no>