peer: replace peer_{start,stop}() with peer_spawn()
The peer is started from the listener, but it is responsible for
stopping itself. Make this a bit clearer by dropping peer_stop()
and renaming peer_start() to peer_spawn().

Also peer_dispatch() is removed, and the affected tests commented
out. It makes no sense to try to do unit tests for the peer, instead
these tests should all be turned into connection tests.

Signed-off-by: Tom Gundersen <teg@jklm.no>