peer: handle duplicate reply serials
In case a message is sent between a pair of peers with a reply serial
that has already been used, it should result in an error, and the
message being dropped.

In order to make the error handling simpler, move queue_message() to
driver.c, where now all the message handling resides.

Signed-off-by: Tom Gundersen <teg@jklm.no>