peer: let the peer own the socket fd
Currently, the socket fd of a peer is owned by the embedded Socket
object (it is responsible for closing the socket when it is
deinitialized). Change this to let the Peer object itself own it
and close it as part of peer_free(). Rename peer_new() to
peer_new_with_fd() to reflect this.

Signed-off-by: Tom Gundersen <teg@jklm.no>