broker: add controller implementation
This is the interface the parent process uses to control the broker.
For now only AddListener(int fd) is supported. Eventually, we want
AddListener() to take a policy attached to each listening fd, and
we want to be able to update said policies and remove listeners
alltogether. Moreover, we want NameSetActivatable() to mark given
names as activatable (or not), and we want a NameActivate signal
to send back to the parent indicating that a given name should be
activated.

Signed-off-by: Tom Gundersen <teg@jklm.no>