controller: complete org.bus1.DBus.Controller API
The API has the following methods:

 AddName(PATH, NAME, UID)
this takes an object path chosen by the caller (must start with /org/bus1/DBus/Name/),
a name that should be activatable and the uid that messages to the name should be
accounted on until the name is activated.

 AddListener(PATH, FD)
this takes an object path chosen by the caller (must start with /org/bus1/DBus/Listener/),
a file descriptor that should be listened on for new connections, and in the future a policy
description for all connection created over this listener.

Signed-off-by: Tom Gundersen <teg@jklm.no>