match: make match.c independent
In particular, do not require a peer object. Instead, introduce a
new MatchOwner, which contains the lookup tree for matches, currently
found in Peer. The MatchOwner object should be embedded in the
corresponding Peer object, allowing the Peer to be accessed via
c_container_of().

In particular, this allows us to do unit-tests for match.c without
pulling in anything else.

Signed-off-by: Tom Gundersen <teg@jklm.no>