tree: fix incorrect use of tree teardown
When we tear down a tree of objects, we better not use postorder
traversal, since it is not stable regarding tree modifications. Fix it
and simply use the root element for now.

Preferably, we would do postorder traversal and not touch the elements.
But that needs some contract with the object that is torn down. So use
the simple way for now and leave it open for future modifications.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>