driver: add name accounting
Update the name queueing to always distinguishing between adding a
new ownership link, and updating an existing one.

Based on this, only charge the quota when actually adding an ownership
relationship. Currently, this means that in case the quota has been
reached, we refuse to update existing names, even though this could
have been allowed. However, we are not interested in making this work,
if the quota has been reached there is already a problem.

Signed-off-by: Tom Gundersen <teg@jklm.no>