drive/match: return dbus error on all invalid match rules
Currently, we consider match rules on unique names that don't exist
to be invalid. This is not strictly in line with the spec, so most
likely we will have to reconsider that. However, at least for now
return the error on the bus properly, rather than consider it a
fatal error.

Signed-off-by: Tom Gundersen <teg@jklm.no>