match: fix tokenizer
Bring the tokenizer in line with what the dbus daemon does. The spec
is not exhaustive, so let's follow the daemon.

A match string is a set of key/value pairs split at (unquoted) commas.
Trailing whitespace and/or equal signs not part of a value is stripped
from the end of the match string. Leading whitespace and/or equal signs
are stripped from keys, and so is trailing whitespace. Keys and values
are delimited by an equal sign. Values are not stripped, but they obey
the quoting logic from the dbus spec.

Signed-off-by: Tom Gundersen <teg@jklm.no>