driver: forward UpdateActivationEnvironment to controller
Ideally, anyone wishing to update the activation enviornemnt would talk to the
activator directly. However, for compatibility with the user bus, we need to
support forwarding such requests.

Currently this does not do any access checks, but it probably should do
something similar to what the dbus daemon does, and not only rely on
policy.

Signed-off-by: Tom Gundersen <teg@jklm.no>