match: track the number of eavesdrop rules
Messages with a destination can only ever be broadcast to eavesdrop
rules. As iterating all matches is potentially expensive, track whether
or not there are any eavesdrop rules installed, and if not, simply
shortcut the matching of unicast messages as nothing can possibly match.

This allows us to drop the global eavesdrop count from the bus object
in a follow-up patch without any adverse effect. In fact, the more
fine-grained we can make this the better.

Signed-off-by: Tom Gundersen <teg@jklm.no>