driver/matches: install black-hole sender matches
When sender matches match on a unique name that is not on the bus,
and never will be, still install it, but don't link it anywhere.

This is needed for backwards compatibility, as installing/removing
such matches should still work as if the peer was there.

We still need to come up with a solution for sender matches for
unique names that may still appear, currently this still fails.

Signed-off-by: Tom Gundersen <teg@jklm.no>