match: allow global sender matches on unique name
So far, the sender (or lack thereof) was implicit in which name
registry a match was installed. This was necessary to deal with
well-known-name matching. However, not strictly speaking necessary
for unique-name maching. These could all be kept in a global list
and the sender id kept in the filter.

The only reason not to do this, and rather install the matches
in the correpsonding sender peer was for performance.

However, for the sake of completeness with respect to the dbus
daemon, we should not fail on matches that match on a unique name
that does not exist (and hence there is no peer to install the
match rule into).

In the case of unique names that have existed, but been disconnected
we simply leave the rule unlinked, as there is no way it can ever
be triggered (unique names are not reused). However, we did not yet
deal with the case of a match on a unique name that may still
appear in the future.

This patch solves that case by installing such rules in the
wildcard registry and support explicit matching by sender
id, if it is specified.

Signed-off-by: Tom Gundersen <teg@jklm.no>