driver: return UID/PID of the controller when the driver is queried
To be compatible with the dbus daemon, we must return credentials also
for the driver. Either we could return the ones for the broker or for the
controller, which in the case of the dbus daemon are the same.

It seems to me that it is the controller that is the right choice, as the
broker itself never acts on the system, appart from through the controller,
when any action is done under the controllers credentials. Moreover, the
credentials of the broker could be queried from the socket directly.

Signed-off-by: Tom Gundersen <teg@jklm.no>