driver: return an error in case of missing Hello() before disconnecting
The dbus daemon does not appear to disconnect the client at all (rather
a following call to Hello() appears to succeed). Move at least a step
in that direction by returning an erro message, but still disconnecting.

The previous commit was important in this regard as it ensures that the
error message is correctly sent to the socket rather than simply dropping
all the pending data.

One could imagine this proving useful in case a client tries to do the
1-to-1 protocol with the daemon (which does not require Hello()), so
returning a proper error rather than a hard disconnect would make
debugging a lot more straight-forward.

Signed-off-by: Tom Gundersen <teg@jklm.no>