driver: unify disconnect handling
Treat all messages before the first Hello() and after BecomeMonitor()
the same: return an error and gracefully shutdown the socket.

No error is returned on invalid message, but the socket is still flushed
and shutdown. It probably makes no sense to try to return an error to
someone not able to marshal a valid message. Though, this might need to
be revisited.

Signed-off-by: Tom Gundersen <teg@jklm.no>