socket: treat invalid messages as EOF, not RESET
For the sake of consistency, react to marshalling errors with a
shutdown, not as a disconnect. This means that no further messages
are read from the socket, but the output buffer is still flushed
to the socket and hung up gracefully.

This fits with how we treat protocol violations in the driver: all
protocol violations cause a graceful shutdown, and violations that
are not marshalling errors also get an error reply.

Signed-off-by: Tom Gundersen <teg@jklm.no>