socket: only discard output on socket_deinit()
The socket object should, to the extent possible, be an extension
of the kernel. Short of freeing the underlying object, nothing should
ever cause socket buffers successfully queued on a socket to be
discarded.

socket_close() now simply disables further dequeueing from the socket
and marks it as shutdown. That means that any output buffers will
still be flushed to the kernel socket, before shutdown() is called.

The old behavior can be restored by simply calling socket_deinit()
immediately after socket_close(), which will drop any pending socket
buffers.

However, the behavoir one probably wants is to wait for socket_dequeue()
to return SOCKET_E_RESET, and only then free the peer (and hence the
socket).

Signed-off-by: Tom Gundersen <teg@jklm.no>