driver/StartServiceByName: only reply with success once the name has been requested
We were passing on the activation request and immediately returning SUCCESS. This
is wrong, and gives the wrong ordering, as the client may rely on the name having
an owner once they receive SUCCESS from this call.

Instead, remember the requst in the activation object, and only answer once the
name has been acquired.

This should fix issue #9.

Signed-off-by: Tom Gundersen <teg@jklm.no>