driver/GetConnectionSELinuxSecurityContext: fix output signature
The output is a byte array, not a bool array. The type character
for byte is 'y', not 'b'. This can only be triggered if the tests
are run with SELinux enabled, so would be good if we could get
a CI that supports that.

The same applies to AdtAudit, but that is dead code anyway.

Should fix issue #10.

Signed-off-by: Tom Gundersen <teg@jklm.no>