controller: add Name.Reset() method
This should be used in case starting a service failed, so pending requests should
be dropped. But future requests should still try to restart the service. Unlike
Release() which removes the name for good.

This is not hooked up from the launcher, but it should listen to systemd signals
and forward activation errors as Name.Reset().

Signed-off-by: Tom Gundersen <teg@jklm.no>