user: allow multiple charges on the same charge object
A UserCharge object is used to track a given charge, allow
the charge to be incremented, as long as it operates on the same
user and actor.

This would be useful if the charge is attached to a buffer, and
we charge it each time more of the buffer is consumed. Under
the assumption that only one user/actor pair can operate on the
given buffer.

Signed-off-by: Tom Gundersen <teg@jklm.no>