connection+socket+peer: rework error/shutdown handling
Socket
------

Any message successfully queued on a socket will be flushed to the
kernel, unless the socket itself is deinitalized early. In
particular, socket_{shutdown,close}() only dissalows future queueing,
but pending socket buffers will still be flushed to the kernel.

Dequeueing a message may return EOF, which means that there was nothing
to dequeue, and there never will be, or it may return RESET, which means
that additionally any further queueing will be discarded. However, in
both cases, the socket may still have outgoing data to flush to the kernel.

If the socket encounters protocol errors when reading/parsing messages
this will be treated like RESET (so the data will be discarded and no
future queue/dequeue will succeed). Note that this is all ordered
correctly: all received messages before the protocol errors are dequeued
correctly, but none thereafter; and all messages queued before the
protocol error are flushed correctly, but none thereafter. No error
will be returned to the remote side as a result of the error, appart
from the disconnect itself (which is ordered as if it was an error
message).

Lastly, a socket is considered running as long as it may still be
possible to (de)queue messages on it, or it has data still to flush
to the kernel. In particular, if a message is not running it can be
deinitialized without losing anything.

Connection
----------

The connection works analogously to the socket, though it may
detect additional protocol errors due to handling SASL.

Peer
----

The peer also works analogously with respect to protocol errors,
with the additional caveat that high-level protocol errors (i.e.,
ones that are not an error when seen as a 1-to-1 connection, but
is an error on the bus) may optionally return an error message
before closing the connection.

When the peer receives CONNECTION_E_RESET, it means the remote end
closed its socket, and the peer should be disconnected.
CONNECTION_E_EOF means the same, but in this case the remote end
may still be able to receive messages so connection_shutdown()
should be called once the request to disconnect has been processed.
This allows the remote end to receive any final messages, such as
NameLost for its unique name.

It is worth noting that in both these cases there may still be
pending buffers in the peer's outgoing queue, and only when the
connection is no longer running should the peer be destroyed.

Signed-off-by: Tom Gundersen <teg@jklm.no>