socket: apply quotas to outgoing buffers
Any message queued on a socket is accounted. There are two cases
to consider:

 * unsolicited messages
These are method calls and directed signals (the latter to be revised).
If such a message fails duo to resource exhaustion an error message
is returned to the sender.

 * solicited messages
These are method replies, error replies, or broadcasts. If such a message
fails due to resource exhaustion the receiver is disconnected.

Put together, this means that no messages are silently dropped. Either
an error is returned or the responsible peer is disconnected.

Signed-off-by: Tom Gundersen <teg@jklm.no>