user/quota: the user itself should not be subject to quotas
A user should be allowed to freely consume as much of its own
resources it desires. Only when consuming resources on behalf
of other users are quotas applied to divide the resources
fairly.

This is already how we deal with maches and names, this patch
extends that logic to also apply to bytes and fds by
distinguishing the case when a user is consuming its own
resources from the case it is consuming resources on behalf of
someone else.

Signed-off-by: Tom Gundersen <teg@jklm.no>