util/dispatch: drop EPOLLRDHUP handling
The EPOLLRDHUP notification is only useful if we 'guess' queue state
from short reads. This is not possible for AF_UNIX sockets, since FD
passing might cut the stream at any point.

Therefore, we can never deduce queue state from short reads, but must
rather rely on EOF to be signalled. EPOLLRDHUP is thus of little use
for AF_UNIX in our edge-triggered dispatcher.

Drop the automatic RDHUP handling for now. If we ever have a use-case,
we can add it again.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>