util/dispatch: never handle events twice in a single round
Rather than moving `handled` events into a @processed list, we do the
inverse: We fetch all events into a @todo list and process them. This
prevents an event from deselecting and reselecting itself or another
event, and thus ending up on the ready-list again, thus possibly
starving the event-loop in case of insufficient handling.

This is in no way severe, since so far we require callbacks to do
forward progress at all time. However, it does feel a lot nicer if we
don't leave this small gap open.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>