user: make user accounting slot-agnostic
Rather than hard-coding all the counters in our user database, make it
a dynamically allocated array. The runtime costs are the same, but we
no longer hard-code all the single counters. Adding and removing
counters is now a lot simpler, and user.[ch] is completely generic and
can be moved to src/util/

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>