metrics: only keep one global metrics object
This measures the time it takes to dispatch a message, not including
any queueing/dequeueing/send/recv, just the time spent in the driver.

At the moment the metrics are not exposed in any way, but we may
in the future add a 'verbose' mode which prints this information on
exit, or possibly a dbus call to query the information.

Signed-off-by: Tom Gundersen <teg@jklm.no>