policy/ownership: add name ownership policy to each peer
Each peer will have three types of policies: one per name, one per name-prefix
and a wildcard one. When a certain query is made a decision is returned which
indiactes if the ownership is permitted or denied together with a priority for
the decision.

When a name is requested a policy decision is queried for that specific name,
for each of its prefixes and for the wildcard name. The highest priority
decision takes precedence.

This is not yet instantiated in any way, so the default 'allow all' policy
is applied.

Signed-off-by: Tom Gundersen <teg@jklm.no>