policy/connection: add connection policy to the bus
The connection policy determines which users can connect to the bus. Though
note that the decision depends on the auxillary groups of the user, which is
resolved at conneciton time and may change between connection attempts.

The bus maintains four types of policies, one per uid, one per gid and a
wildcard policy for both. As with name ownership policies, each of these may
be queried for a decision, which will return whether or not the connection
should be allowed, and the priority of the decision.

When a connection is attempted, the policy for the connection uid is queried
as well as the policy for each of its auxillary gids as well as the wildcard
policies for both uids and gids. The decision with the highest priority takes
precedence.

This is not yet instantiated in any way, so the default 'allow all' policy
is applied. Moreover, the hookup to resolve auxillary groups is still
missing.

Signed-off-by: Tom Gundersen <teg@jklm.no>