policy/transmission: add send and receive policies to each peer (and the driver)
The transmission policies determine which peers can send and receive messages
to and from eachother. The set of policies for a given peer is determined at
connection time based on its uid, and stays fixed after that. Note, however,
that what policies a uid resolves to depends on its auxillary groups, and
whether or not it is marked as being 'at_console', at the time of connection.
In other words, it is possible for different peers of the same uid to have
different policies.

Each peers maintains two transmission policies; a send policy and a receive
policy (in subsequent patches these will come in eavesdropping and
non-eavesdropping versions). For a given message to be allowed to be
transmitted it must be allowed to be sent according to the send policy of
the sender and to be allowed to be received according to the receive policy
of the receiver.

When a transmission policy is queried, it takes as input some metadata about
the message being sent, along with a 'subject', which is the other peer
taking part in the transaction (the receiver for a send policy, or the sender
for a receive policy). The policy is queried once for each name the subject
owns and once for the wildcard name. As with the other policies, the returned
result with the highest priority takes precedence.

This is not yet instantiated in any way, so the default 'allow all' policy
is applied.

Signed-off-by: Tom Gundersen <teg@jklm.no>