policy: add Policy and PolicyRegistry objects
A Policy object is simply a collection of Ownership and Transmission
policies, to be embedded in each Peer.

The PolicyRegistry holds a collection of Policy objects from which
each per-Peer Policy object is instantiated. It also contains the
ConnectionPolicy determining which uid's may or may not connect.

Ideally, the PolicyRegistry would simply hold a set of policies
indexed by uid's and each peer would simply take a reference on the
object corresponding to its uid, but as long as we have group and
at_console policies, we need to give each peer its own instantiation
of its policy, as it may differ between connectes (even when the uid
does not).

Signed-off-by: Tom Gundersen <teg@jklm.no>