policy/parser: parse the XML policy into a PolicyRegistry
Each allow/deny rule is assigned a priority. We ensure tha the
priority is such that:

  mandatory > at_console > user > group > default

Moreover, for any two rules whose priorities are not given by the
above we take the one appearing later in the file order to have
the highest priority.

This should give the same behavior as the dbus daemon, with the
only exception that they say the order between different groups
are arbitrary, while we still respecte the file order.

This is implemented by having a counter that is incremented for
each rule we encounter as we parse the files, as well as a base
offset correpsonding to each of the five types of rules. We rely
on the fact that the counter will never overflow, and take the
priority of a rule to be the sum of its counter and the base.

Signed-off-by: Tom Gundersen <teg@jklm.no>