peer/policy: replace explicit connection check with policy instantiation
For now this simply does the same conneciton policy checks, but in
follow-up commits this will also instantiate the actual policy object.

Note that checking for auxillary groups is potentially nasty (can cause
dead-locks in the worst case), so we only do this if there are group
rules installed on the system.

Signed-off-by: Tom Gundersen <teg@jklm.no>