peer: resolve auxillary groups if necessary
Instantiating a policy for a peer may depend on its auxillary groups.
In this case, resolve the groups when creating the peer and pass them
on. This moves the functionality from policy.c.

First try fetching the groups from the socket, before falling back
to resolving them using nss. Resolving groups using nss is problematic
as nss itself may be using dbus, and hence cause a deadlock. However,
for now we fall back to this on kernel without SO_PEERGROUPS support.

Signed-off-by: Tom Gundersen <teg@jklm.no>