peer/getgrouplist: avoid infinite loop on error
getgrouplist(3) returns -1 when a bigger gid array needs to be passed in.
However, it also returns -1 if an internal error occured. The only way to
detect this case, is to observe that the total number of gids needed
did not grow from the previous call. Detect this case and bail out.

Signed-off-by: Tom Gundersen <teg@jklm.no>