bus: make bus_find_peer_by_name() return name information
Rather than hard-coding anything that needs path information for peer
lookups, lets make bus_find_peer_by_name() return that information.

This changes bus_find_peer_by_name() to always return the Name used for
lookup in an optional output argument. This is done regardless whether
the name as an owner or not. With this in place, we can drop 3 places
in driver.c that did peer lookups manually.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>